Create a new page for each tab
