+++ src/components/Dashboard/CreateRoom.generated.ts
@@ -1,4 +1,4 @@
-import * as Types from '../../../graphql';
+import * as Types from 'graphql';
 
 import { gql } from '@apollo/client';
 import * as Apollo from '@apollo/client';
