+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import { Route, Switch } from "react-router";
 import { Home } from "./Home";
-import { Dashboard } from "pages/room/Room";
 import { LoginRouter } from "pages/login";
 import { RoomRouter } from "./room";
 
@@ -11,11 +10,6 @@ export const Pages = () => (
 
     <Route path="/room" render={() => <RoomRouter />} />
     <Route path="/login" render={() => <LoginRouter />} />
-
-    <Route path="/dashboard" render={() => <Dashboard />} />
-    <Route path="/dashboard/controller" />
-    <Route path="/dashboard/rooms" />
-    <Route path="/dashboard/account" />
 
     <Route path="/help-center" />
     <Route path="/help-center/:helpCenterArticleGuid" />
