+++ src/pages/room/ControllerSettingsTab/index.tsx
@@ -1,0 +1,15 @@
+import {
+  Heading,
+  Box,
+} from "@chakra-ui/react";
+
+import { Dashboard } from "components/Dashboard";
+
+export const ControllerSettingsTab = () => {
+  return (
+    <Dashboard>
+      <Box>
+        <Heading> Controller Settings Tab </Heading>
+      </Box>
+    </Dashboard>);
+}
