+++ src/pages/room/PlayersTab/index.tsx
@@ -1,0 +1,15 @@
+import {
+  Heading,
+  Box,
+} from "@chakra-ui/react";
+
+import { Dashboard } from "components/Dashboard";
+
+export const PlayersTab = () => {
+  return (
+    <Dashboard>
+      <Box>
+        <Heading> Players Tab </Heading>
+      </Box>
+    </Dashboard>);
+}
