+++ src/pages/room/StreamTab/index.tsx
@@ -1,0 +1,15 @@
+import {
+  Heading,
+  Box,
+} from "@chakra-ui/react";
+
+import { Dashboard } from "components/Dashboard";
+
+export const StreamTab = () => {
+  return (
+    <Dashboard>
+      <Box>
+        <Heading> Stream Tab </Heading>
+      </Box>
+    </Dashboard>);
+}
